pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "wa-crypt-tools"
keywords = ["whatsapp", "crypt12", "crypt14", "crypt15"]
version = "0.0.8"
authors = [
    {name = "Davide Palma", email = "pypi@davidepalma.it"},
]
description = "Decrypts WhatsApp .crypt12, .crypt14 and .crypt15 files, given the key."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "javaobj-py3 @ git+https://github.com/tcalmant/python-javaobj@fba64e9",
    "pycryptodomex ~= 3.18",
    "protobuf >= 4.24.1,< 4.25.0"
]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",

]

[project.scripts]
wadecrypt = "wa_crypt_tools.wadecrypt:main"
waguess = "wa_crypt_tools.waguess:main"

[project.optional-dependencies]
proto = [
    "protoletariat"
]